feat(ui): move common components to DS-components
