+++ ui/apps/moderation/src/components/common/index.tsx
@@ -1,5 +1,3 @@
 export * from './category-pills';
-export * from './page-buttons';
 export * from './page-header';
-export * from './stepped-action-wrapper';
 export * from './subtitle-renderer';
